# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Add numbers service",
                "operationId": "2",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "first number",
                        "name": "num1",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "second number",
                        "name": "num2",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.AddResponse"
                        }
                    }
                }
            }
        },
        "/bin": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Decimal to binary service",
                "operationId": "4",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number",
                        "name": "num1",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResultResponse"
                        }
                    }
                }
            }
        },
        "/div": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Division service",
                "operationId": "6",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "first number",
                        "name": "num1",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "second number",
                        "name": "num2",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.DivResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Health service",
                "operationId": "7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.GreeterResponse"
                        }
                    }
                }
            }
        },
        "/mul": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Multiply service",
                "operationId": "5",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "first number",
                        "name": "num1",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "second number",
                        "name": "num2",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MulResponse"
                        }
                    }
                }
            }
        },
        "/sub": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Subtract service",
                "operationId": "3",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "first number",
                        "name": "num1",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "second number",
                        "name": "num2",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SubResponse"
                        }
                    }
                }
            }
        },
        "/v1/greeter": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Greeter service",
                "operationId": "1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Input name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.GreeterResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AddResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer"
                }
            }
        },
        "main.DivResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer"
                }
            }
        },
        "main.GreeterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.MulResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer"
                }
            }
        },
        "main.ResultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.SubResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer"
                }
            }
        }
    }
}